git-svn-id: https://bsisvn.mayo.edu/main/projects/g/Genome_GPS/trunk@1096 8f329c14-d232-4f17-8684-7ee34322b8dc

+++ polyphen.sh
@@ -26,24 +26,30 @@ else
     snv_file=$sample.variants.chr$chr.SNV.filter.i.c.vcf
     
     cat $input/$snv_file | awk '$0 !~ /^#/' | awk '{print $1":"$2"\t"$4"/"$5}' > $polyphen/$snv_file.poly
-    
+    num=`cat $polyphen/$snv_file.poly |wc -l `
     ### get the uniport id
-    $perl $pph/bin/mapsnps.pl -A -g $genome_build $polyphen/$snv_file.poly > $polyphen/$snv_file.poly.uniprot
+    if [ $num -gt 0 ]
+    then
+        $perl $pph/bin/mapsnps.pl -A -g $genome_build $polyphen/$snv_file.poly > $polyphen/$snv_file.poly.uniprot
+    else
+        touch $polyphen/$snv_file.poly.uniprot
+    fi
     rm $polyphen/$snv_file.poly
     
-    ### get the prediction
-    uniprot=`awk '{ for(i=1;i<=NF;i++){ if ($i == "spacc") {print i} } }' $polyphen/$snv_file.poly.uniprot`
-    aa1=`awk '{ for(i=1;i<=NF;i++){ if ($i == "aa1") {print i} } }' $polyphen/$snv_file.poly.uniprot`
-    aa2=`awk '{ for(i=1;i<=NF;i++){ if ($i == "aa2") {print i} } }' $polyphen/$snv_file.poly.uniprot`
-    aapos=`awk '{ for(i=1;i<=NF;i++){ if ($i == "cdnpos") {print i} } }' $polyphen/$snv_file.poly.uniprot`
-    
-    cat $polyphen/$snv_file.poly.uniprot | awk 'NR>1' | awk -v a1=$aa1 -v a2=$aa2  -F '\t' '$a1 !~ $a2' |  awk -v uni=$uniprot 'length($uni)>1' | awk -v chr=$snp_pos -v uni=$uniprot -v a1=$aa1 -v a2=$aa2 -v pos=$aapos -F '\t' '{print $uni"\t"$pos"\t"$a1"\t"$a2}' > $polyphen/$snv_file.poly.uniprot.in
-    
-    $perl $pph/bin/run_pph.pl -A -g $genome_build -d $polyphen/$sample.$chr $polyphen/$snv_file.poly.uniprot.in > $polyphen/$snv_file.poly.uniprot.in.predict
-    rm $polyphen/$snv_file.poly.uniprot.in
-    
-    ### map chr pos with the prediction
-    
+    num=`cat $polyphen/$snv_file.poly.uniprot | wc -l`
+    if [ $num -gt 0 ]
+    then
+        ### get the prediction
+        uniprot=`awk '{ for(i=1;i<=NF;i++){ if ($i == "spacc") {print i} } }' $polyphen/$snv_file.poly.uniprot`
+        aa1=`awk '{ for(i=1;i<=NF;i++){ if ($i == "aa1") {print i} } }' $polyphen/$snv_file.poly.uniprot`
+        aa2=`awk '{ for(i=1;i<=NF;i++){ if ($i == "aa2") {print i} } }' $polyphen/$snv_file.poly.uniprot`
+        aapos=`awk '{ for(i=1;i<=NF;i++){ if ($i == "cdnpos") {print i} } }' $polyphen/$snv_file.poly.uniprot`
+        cat $polyphen/$snv_file.poly.uniprot | awk 'NR>1' | awk -v a1=$aa1 -v a2=$aa2  -F '\t' '$a1 !~ $a2' |  awk -v uni=$uniprot 'length($uni)>1' | awk -v chr=$snp_pos -v uni=$uniprot -v a1=$aa1 -v a2=$aa2 -v pos=$aapos -F '\t' '{print $uni"\t"$pos"\t"$a1"\t"$a2}' > $polyphen/$snv_file.poly.uniprot.in
+        $perl $pph/bin/run_pph.pl -A -g $genome_build -d $polyphen/$sample.$chr $polyphen/$snv_file.poly.uniprot.in > $polyphen/$snv_file.poly.uniprot.in.predict
+        rm $polyphen/$snv_file.poly.uniprot.in
+    else
+        touch $polyphen/$snv_file.poly.uniprot.in.predict
+    fi    
     perl $script_path/map.polyphen.pl $polyphen/$snv_file.poly.uniprot $polyphen/$snv_file.poly.uniprot.in.predict > $polyphen/$snv_file.poly
     rm $polyphen/$snv_file.poly.uniprot
     rm $polyphen/$snv_file.poly.uniprot.in.predict
